# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API Rumah Sakit",
        "description": "API ini digunakan untuk mengelola data pasien, dokter, konsultasi, dan lainnya.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Server API Lokal"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/laporan": {
            "get": {
                "tags": [
                    "Laporan"
                ],
                "summary": "Menampilkan semua data laporan",
                "operationId": "076e4c0226607e0351c07a8e1ac1893c",
                "responses": {
                    "200": {
                        "description": "Data laporan berhasil diambil"
                    }
                }
            },
            "post": {
                "tags": [
                    "Laporan"
                ],
                "summary": "Menyimpan data laporan baru",
                "operationId": "d97fa4e1a3072819b633afae97511b2e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "periode",
                                    "jumlah_pasien",
                                    "pendapatan"
                                ],
                                "properties": {
                                    "periode": {
                                        "type": "string",
                                        "example": "Januari 2025"
                                    },
                                    "jumlah_pasien": {
                                        "type": "integer",
                                        "example": 100
                                    },
                                    "pendapatan": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 15000000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Data laporan berhasil disimpan"
                    }
                }
            }
        },
        "/api/laporan/{id}": {
            "get": {
                "tags": [
                    "Laporan"
                ],
                "summary": "Menampilkan detail laporan berdasarkan ID",
                "operationId": "c38dac1985b9507737fabf48a36d6580",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID laporan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail laporan ditemukan"
                    },
                    "404": {
                        "description": "Laporan tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Laporan"
                ],
                "summary": "Memperbarui data laporan",
                "operationId": "cf5b58b0489047876ec7e684d8482324",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID laporan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "periode": {
                                        "type": "string",
                                        "example": "Februari 2025"
                                    },
                                    "jumlah_pasien": {
                                        "type": "integer",
                                        "example": 120
                                    },
                                    "pendapatan": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 17000000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data laporan berhasil diperbarui"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Laporan"
                ],
                "summary": "Menghapus data laporan",
                "operationId": "c9e61853f673d1cb6998868e949f0249",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID laporan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Laporan berhasil dihapus"
                    }
                }
            }
        },
        "/api/pasien": {
            "get": {
                "tags": [
                    "Pasien"
                ],
                "summary": "Menampilkan daftar semua pasien",
                "operationId": "getPasien",
                "responses": {
                    "200": {
                        "description": "Berhasil mendapatkan daftar pasien"
                    }
                }
            },
            "post": {
                "tags": [
                    "Pasien"
                ],
                "summary": "Menyimpan data pasien baru",
                "operationId": "storePasien",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama",
                                    "tanggal_lahir",
                                    "jenis_kelamin",
                                    "alamat"
                                ],
                                "properties": {
                                    "nama": {
                                        "type": "string"
                                    },
                                    "tanggal_lahir": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "jenis_kelamin": {
                                        "type": "string"
                                    },
                                    "alamat": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Data pasien berhasil disimpan"
                    }
                }
            }
        },
        "/api/pasien/{id}": {
            "get": {
                "tags": [
                    "Pasien"
                ],
                "summary": "Menampilkan detail pasien berdasarkan ID",
                "operationId": "getDetailPasien",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail pasien ditemukan"
                    },
                    "404": {
                        "description": "Pasien tidak ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Pasien"
                ],
                "summary": "Memperbarui data pasien",
                "operationId": "updatePasien",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama",
                                    "tanggal_lahir",
                                    "jenis_kelamin",
                                    "alamat"
                                ],
                                "properties": {
                                    "nama": {
                                        "type": "string"
                                    },
                                    "tanggal_lahir": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "jenis_kelamin": {
                                        "type": "string"
                                    },
                                    "alamat": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data pasien berhasil diperbarui"
                    },
                    "404": {
                        "description": "Pasien tidak ditemukan"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Pasien"
                ],
                "summary": "Menghapus data pasien",
                "operationId": "deletePasien",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pasien berhasil dihapus"
                    },
                    "404": {
                        "description": "Pasien tidak ditemukan"
                    }
                }
            }
        },
...
    },
...
}